+++ lua/plugins/debug.lua
@@ -1,0 +1,3 @@
+-- Debug Adapter Protocol (DAP) configuration
+-- Re-exports the existing kickstart debug config
+return require 'kickstart.plugins.debug'
