+++ lua/plugins/deltaview.lua
@@ -1,0 +1,11 @@
+return {
+  {
+    'kokusenz/deltaview.nvim',
+    opts = {
+      keyconfig = {
+        dv_toggle_keybind = '<leader>dd',
+      },
+      use_nerdfonts = false,
+    },
+  },
+}
