+++ lua/plugins/git.lua
@@ -1,0 +1,19 @@
+-- Git integration plugins
+return {
+  -- Git commands in nvim
+  'tpope/vim-fugitive',
+
+  -- Git signs in the gutter
+  {
+    'lewis6991/gitsigns.nvim',
+    opts = {
+      signs = {
+        add = { text = '+' },
+        change = { text = '~' },
+        delete = { text = '_' },
+        topdelete = { text = '‾' },
+        changedelete = { text = '~' },
+      },
+    },
+  },
+}
